✨ feat: add command pattern support for processors
Introduced support for the command pattern in processors through the `command_class` property. If set, the processor instantiates a command object, providing data isolation and cleaner design. Added the `StatefulChainedProcessor` and `StatefulChainedPipeline` as command-pattern versions of the chained processor. All related code is in the `partials/with_processor_commands` folder. Also included pytest tests for the new pipeline and processor.